move some methods to TsBlockUtil
